+++ tests/bad_servers/bad_init_args2.py
@@ -3,5 +3,5 @@ from msl.loadlib import Server32
 
 class BadInitArgs2(Server32):
 
-    def __init__(self, host, port, quiet, extra, **kwargs):
+    def __init__(self, host, port, extra, **kwargs):
         pass