+++ tests/bad_servers/unexpected_error2.py
@@ -3,8 +3,6 @@ from msl.loadlib import Server32
 
 class UnexpectedError(Server32):
 
-    def __init__(self, host, port, quiet, **kwargs):
-        # include the 'quiet' positional argument
-
+    def __init__(self, host, port, **kwargs):
         # any error would be fine
         x = 1 + 'hello'